ANNDEA v3: Debugging slow CNN training...

+++ Code/Utilities/U_ML.py
@@ -954,7 +954,8 @@ def LoadRenderImages(Seeds,StartSeed,EndSeed,num_classes=2):
     NewSeeds=Seeds[StartSeed-1:min(EndSeed,len(Seeds))]
     ImagesY=np.empty([len(NewSeeds),1])
 
-    ImagesX=np.empty([len(NewSeeds),NewSeeds[0].H,NewSeeds[0].W,NewSeeds[0].L],dtype=np.uint8)
+    #ImagesX=np.empty([len(NewSeeds),NewSeeds[0].H,NewSeeds[0].W,NewSeeds[0].L],dtype=np.uint8)
+    ImagesX=np.empty([len(NewSeeds),NewSeeds[0].H,NewSeeds[0].W,NewSeeds[0].L],dtype=np.float16)
     for im in range(len(NewSeeds)):
         if num_classes>1:
             if hasattr(NewSeeds[im],'Label'):
